first draft for a history plot to show parameter values over time
i prepare the plot data in directly in the view because the function that prepares all the button for navigation has every test with the same name of the same project already looked up at some point. i guess performance wise its useful to also prepare the plot data here. all the scripting and chart stuff is pretty WIP right now and hardcoded, as to show what is happening. the graph itself is technically working, although the position of the graph itself is obviously just for debugging.