Improve handling of settings
The `rest` sub-directory has been deleted. Instead a `settings` sub-directory was added to handle the settings of the whole application. The settings package was divided into two parts, a `development.py` file for development settings a`local.py` settings file that can be adjusted for settings used in production. Both setting configurations are based on the settings defined in `common.py`. The settings package will first try to load a `local.py` and fall back to the `development.py` settings when the local settings are not availble.

The file `local.py.example` is supposed to provide an entry point that a user can adjust to provide settings for a production instance of the web app.