Improve test execution speed
In the Webhook tests, when we prepare the test data, the webhooks for the models are kicking in already, which results in POST request getting generated to invalid URLs.

These invalid requests block the test execution. In this commit we replace the `_submit_webhook_request` function to a no-op, so that the requests are not being posted.