Update columns displayed for submissions in project details
We do only display the submission "created" date and remove the "last updated" date, since in pretty much all cases, these two dates will be the same anyways. (the last updated date is not updated when a test for the submission is updated, but only when the submission itself is being updated, which pretty much never happens).

Instead we show the status for each submission and move the submission ID into second place after the status.