Do not allow unrelated model instances
Change the models to not allow "null" relations for ProjectSubmissionProperties, ReferenceSets, Submissions, TestReferences and TestResults.

In does not make sense to have objects of those models in the data-base that do not belong to any of their specified "parent" objects.

The migration will simply delete any unrelated objects from the database.