Use ModelSerializers as base classes for serializers
The `ModelSerializer` base class allows us to simplify the serializer classes a lot, since we do not need to duplicate most data from the model fields. Additionally it is more maintainable, since changes in the model fields will automatically be applied to the sterilizers.

As a side effect relations are handled by `PrimaryKeyRelatedField` now, which simplifies most of the validation routines.

One change worth mentioning here is that it is no longer possible to specify projects as either id, slug or unique name in the serializers anymore. Only the id is accepted now. This is not a real problem, since specifying the id was no longer done in the users payload anymore, but is handled by the URL of the POST requests.