Improve representation of nd-arrays
They are now printed as small tables. For vector arrays (n=1) we print a table with a single column, matrices (n=2) are printed regular tables with the correct amount of rows and columns. Tensors of order 3 (n=3) are printed as multiple tables and Tensors of any other order (n>=4) are just printed as a list with a values, with an additional header that states the tensors order and it's shape.