Redirect the root page to the projects overview page
The front-page is currently designed to show the most recent submissions for all projects, which might not be very helpful if a new user wants do discover the functionality of DTF (e.g. there is no button to get to the projects page).

Redirecting to the projects overview page should be more useful until we get some better homepage, since one should be able to discover everything else staring from the projects overview page.