Add filter to create test result measurement table entries
This new filter translates a test-result/test-reference value to its HTML representation and puts that into a link if the test-result/test-reference has a test source id available.
Additionally this handles the case where either the complete entry or the value is None and prints those as "N/A" text (while still keeping the link, if the None value has a source).