Change behavior when updating references to non-existing value
This commit changes what happens when we want to update a test result entry that had a reference, but no current value is available.

Before this commit, the value was removed from the test references completely. After this commit, it will still be stored as `None` value in the references.

This allows to keep track of references that have been available in a previous state, but have been removed in a newer submission.