Add support to create unique submissions
We add the possibility to pass a "unique_key" in the query params of a POST request to the submissions endpoint.
The value for that key will then be looked up in the submissions `info` dictionary, and if any other submission with the same key value pair does already exit, a 409 (conflict) response will be send and no new submission will be created.