Improve links in table rows
Instead of the whole row acting as a link, we do now make only specific table items links.
This way we can use real "<a>" links, which give better control in the browser (e.g. we can still select text in the row and copy it and we can right-click the links to select things like "open in new tab").